page-analysis-service/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
sentence-transformers
keybert
redis
python-dotenv
pydantic
scikit-learn
prometheus-fastapi-instrumentator
